+++ cmd/version_generated.go
@@ -26,6 +26,7 @@ func VersionCommand() *cobra.Command {
 			return PrepareConfig(cmd, &metadata, "version", &myVersionOptions, config.OpenPiperFile)
 		},
 		RunE: func(cmd *cobra.Command, args []string) error {
+
 			return version(myVersionOptions)
 		},
 	}
