+++ pkg/generator/helper/resources_test.go
@@ -1,4 +1,4 @@
-package config
+package helper
 
 import (
 	"fmt"
